Backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
psycopg2-binary
python-dotenv
Celery
redis
